requirements.txt
langchain-openai>=1.0.1
langchain-core>=1.0.0
langgraph>=1.0.1
openai>=2.6.0
python-dotenv>=1.1.1

